feat(#77): add support for future namespace configuration
