+++ lib/src/main/java/com/telemetrydeck/sdk/TelemetryApiClientFactory.kt
@@ -1,0 +1,12 @@
+package com.telemetrydeck.sdk
+
+import java.net.URL
+
+interface TelemetryApiClientFactory {
+    fun create(
+        apiBaseURL: URL,
+        showDebugLogs: Boolean,
+        namespace: String?,
+        logger: DebugLogger?
+    ): TelemetryApiClient
+}
