+++ lib/src/main/java/com/telemetrydeck/sdk/TelemetryClientFactory.kt
@@ -1,0 +1,19 @@
+package com.telemetrydeck.sdk
+
+import java.net.URL
+
+internal class TelemetryClientFactory: TelemetryApiClientFactory {
+    override fun create(
+        apiBaseURL: URL,
+        showDebugLogs: Boolean,
+        namespace: String?,
+        logger: DebugLogger?
+    ): TelemetryApiClient {
+        return TelemetryClient(
+            apiBaseURL,
+            showDebugLogs,
+            namespace,
+            logger
+        )
+    }
+}
